requirements.txt
click==7.1.2
ecdsa==0.14.1
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
mysqlclient==2.0.3
psycopg2==2.8.6
pyasn1==0.4.8
pydantic==1.8.2
PyMySQL==1.0.2
python-decouple==3.4
python-jose==3.2.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
